+++ packages/manager/.changeset/pr-10414-tech-stories-1714167805545.md
@@ -1,0 +1,5 @@
+---
+"@linode/manager": Tech Stories
+---
+
+Query Key Factory for Volumes ([#10414](https://github.com/linode/manager/pull/10414))
