+++ packages/manager/src/features/Linodes/LinodesDetail/LinodesDetailHeader/UpgradeVolumesDialog.tsx
@@ -9,8 +9,10 @@ import { Typography } from 'src/components/Typography';
 import { VolumeUpgradeCopy } from 'src/features/Volumes/UpgradeVolumeDialog';
 import { getUpgradeableVolumeIds } from 'src/features/Volumes/utils';
 import { useNotificationsQuery } from 'src/queries/account/notifications';
-import { useLinodeVolumesQuery } from 'src/queries/volumes';
-import { useVolumesMigrateMutation } from 'src/queries/volumesMigrations';
+import {
+  useLinodeVolumesQuery,
+  useVolumesMigrateMutation,
+} from 'src/queries/volumes/volumes';
 import { getAPIErrorOrDefault } from 'src/utilities/errorUtils';
 
 import type { Linode } from '@linode/api-v4';
