+++ packages/manager/src/features/Linodes/MigrateLinode/CautionNotice.tsx
@@ -7,7 +7,7 @@ import { Link } from 'src/components/Link';
 import { Notice } from 'src/components/Notice/Notice';
 import { Typography } from 'src/components/Typography';
 import { API_MAX_PAGE_SIZE } from 'src/constants';
-import { useLinodeVolumesQuery } from 'src/queries/volumes';
+import { useLinodeVolumesQuery } from 'src/queries/volumes/volumes';
 
 interface Props {
   edgeRegionWarning?: string;
