+++ packages/manager/src/features/Volumes/AttachVolumeDrawer.tsx
@@ -14,7 +14,7 @@ import { LinodeSelect } from 'src/features/Linodes/LinodeSelect/LinodeSelect';
 import { useEventsPollingActions } from 'src/queries/events/events';
 import { useAllLinodeConfigsQuery } from 'src/queries/linodes/configs';
 import { useGrants } from 'src/queries/profile';
-import { useAttachVolumeMutation } from 'src/queries/volumes';
+import { useAttachVolumeMutation } from 'src/queries/volumes/volumes';
 import { getAPIErrorFor } from 'src/utilities/getAPIErrorFor';
 
 interface Props {
