+++ packages/manager/src/features/Volumes/EditVolumeDrawer.tsx
@@ -9,7 +9,7 @@ import { Notice } from 'src/components/Notice/Notice';
 import { TagsInput } from 'src/components/TagsInput/TagsInput';
 import { TextField } from 'src/components/TextField';
 import { useGrants } from 'src/queries/profile';
-import { useUpdateVolumeMutation } from 'src/queries/volumes';
+import { useUpdateVolumeMutation } from 'src/queries/volumes/volumes';
 import {
   handleFieldErrors,
   handleGeneralErrors,
