+++ packages/manager/src/features/Volumes/VolumeDrawer/LinodeVolumeAttachForm.tsx
@@ -9,7 +9,7 @@ import { ActionsPanel } from 'src/components/ActionsPanel/ActionsPanel';
 import { Notice } from 'src/components/Notice/Notice';
 import { useEventsPollingActions } from 'src/queries/events/events';
 import { useGrants } from 'src/queries/profile';
-import { useAttachVolumeMutation } from 'src/queries/volumes';
+import { useAttachVolumeMutation } from 'src/queries/volumes/volumes';
 import {
   handleFieldErrors,
   handleGeneralErrors,
