+++ packages/manager/src/hooks/useEventHandlers.ts
@@ -5,17 +5,17 @@ import { databaseEventsHandler } from 'src/queries/databases';
 import { domainEventsHandler } from 'src/queries/domains';
 import { firewallEventsHandler } from 'src/queries/firewalls';
 import { imageEventsHandler } from 'src/queries/images';
-import { linodeEventsHandler } from 'src/queries/linodes/events';
 import { diskEventHandler } from 'src/queries/linodes/events';
+import { linodeEventsHandler } from 'src/queries/linodes/events';
 import { nodebalanacerEventHandler } from 'src/queries/nodebalancers';
 import { sshKeyEventHandler } from 'src/queries/profile';
+import { stackScriptEventHandler } from 'src/queries/stackscripts';
 import { supportTicketEventHandler } from 'src/queries/support';
 import { tokenEventHandler } from 'src/queries/tokens';
-import { volumeEventsHandler } from 'src/queries/volumes';
+import { volumeEventsHandler } from 'src/queries/volumes/events';
 
 import type { Event } from '@linode/api-v4';
 import type { QueryClient } from '@tanstack/react-query';
-import { stackScriptEventHandler } from 'src/queries/stackscripts';
 
 export interface EventHandlerData {
   event: Event;
